# swagger.yaml
swagger: "2.0"
info:
  version: "2.0.0"
  title: "Hotel Audit System API"
  description: "Please consider that resources with a lock require a valid secretKey query parameter on the /v1 path (\"v1_auth\") and a valid JWT access token (\"v2_auth_query\" or \"v2_auth_header\") on the /v2 path."
  contact:
    email: "martin.biermann@hrs.com"
host: "api.hotel-audit.hrs.com"
...
schemes:
- "https"
paths:
  /v1/hotel-status:
    get:
      tags:
      - "v1"
      deprecated: true
      summary: ""
      description: ""
      operationId: "getHotelStatus"
      produces:
      - "application/json"
      parameters:
      - name: "hkeys"
        in: "query"
        description: "Comma-separated list of hkeys"
        required: true
        type: "string"
      responses:
        "200":
          description: "Successful operation"
        "401":
          description: "Unauthorized access"
        "403":
          description: "Forbidden"
        "500":
          description: "Server error"
      security:
      - v1_auth: []
...
  /v1/invitations/hotels:
    get:
      tags:
      - "v1"
      deprecated: true
      summary: "Pagination on Clean & Safe invitations"
      description: "Paginated listing of Clean & Safe invitations"
      operationId: "getHotelInvitations"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "query"
        description: "Number of page"
        required: true
        type: "integer"
        minimum: 1
        default: 1
      - name: "size"
        in: "query"
        description: "Number of results per page"
        required: true
        type: "integer"
        minimum: 1
        maximum: 500
      responses:
        "200":
          description: "Successful operation"
        "401":
          description: "Unauthorized access"
        "403":
          description: "Forbidden"
        "500":
          description: "Server error"
      security:
      - v1_auth: []
  /v1/audits:
    get:
      tags:
      - "v1"
      deprecated: true
      summary: ""
      description: ""
      operationId: "getAudits"
      produces:
      - "application/json"
      parameters:
      - name: "hkeys"
        in: "query"
        description: "Comma-separated list of hkeys"
        required: true
        type: "string"
      responses:
        "200":
          description: "Successful operation"
        "401":
          description: "Unauthorized access"
        "403":
          description: "Forbidden"
        "500":
          description: "Server error"
      security:
      - v1_auth: []
...
  /v1/cs_cert/{code}:
    get:
      tags:
      - "v1"
      summary: "Clean & Safe Certificate"
      description: "Clean & Safe Certificate"
      operationId: "getCleanSafeCert"
      produces:
      - "text/html"
      parameters:
      - name: "code"
        in: "path"
        description: "Request code of certificate"
        type: "string"
        required: true
      responses:
        "200":
          description: "Successful operation"
        "403":
          description: "Forbidden"
        "500":
          description: "Server error"
  /v1/cs_widget/{type}/{code}:
    get:
      tags:
      - "v1"
      summary: "Clean & Safe Widget"
      description: "Clean & Safe Widget"
      operationId: "getCleanSafeWidget"

      produces:
      - "image/png"
      parameters:
      - name: "type"
        in: "path"
        description: "Type of widget"
        type: "string"
        required: true
      - name: "code"
        in: "path"
        description: "Request code of widget"
        type: "string"
        required: true
      responses:
        "200":
          description: "Successful operation"
        "403":
          description: "Forbidden"
        "500":
          description: "Server error"
  /v1/cs_sticker/{type}/{code}:
    get:
      tags:
      - "v1"
      summary: "Clean & Safe Sticker"
      description: "Clean & Safe Sticker"
      operationId: "getCleanSafeSticker"

      produces:
      - "application/pdf"
      parameters:
      - name: "type"
        in: "path"
        description: "Type of widget"
        type: "string"
        required: true
      - name: "code"
        in: "path"
        description: "Request code of widget"
        type: "string"
        required: true
      responses:
        "200":
          description: "Successful operation"
        "403":
          description: "Forbidden"
        "500":
          description: "Server error"
  /v1/cs_widget_link/{type}/{code}:
    get:
      tags:
      - "v1"
      summary: "Clean & Safe Widget Link Redirect"
      description: "Clean & Safe Link Redirect"
      operationId: "getCleanSafeLink"

      parameters:
      - name: "type"
        in: "path"
        description: "Type of widget"
        type: "string"
        required: true
      - name: "code"
        in: "path"
        description: "Request code of widget"
        type: "string"
        required: true
      responses:
        "302":
          description: "Successful operation"
        "403":
          description: "Forbidden"
        "500":
          description: "Server error"
...
  /v2/hotel-status:
    get:
      tags:
      - "v2"
      summary: ""
      description: ""
      operationId: "getHotelStatusV2"
      produces:
      - "application/json"
      parameters:
      - name: "hkeys"
        in: "query"
        description: "Comma-separated list of hkeys"
        required: true
        type: "string"
      - name: "backfill"
        in: "query"
        description: "***For internal and educational purposes only - should not be used in public production environments.***\nOptional flag to provide calculated Green Stay record based on similar hotels, if no actual record exists for the hotel."
        required: false
        type: "boolean"
      responses:
        "200":
          description: "Successful operation"
        "401":
          description: "Unauthorized access"
        "403":
          description: "Forbidden"
        "500":
          description: "Server error"
      security:
      - v2_auth_query: []
      - v2_auth_header: []
  /v2/invitations/hotels:
    get:
      tags:
      - "v2"
      summary: "Pagination on Clean & Safe invitations"
      description: "Paginated listing of Clean & Safe invitations"
      operationId: "getHotelInvitationsV2"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "query"
        description: "Number of page"
        required: true
        type: "integer"
        minimum: 1
        default: 1
      - name: "size"
        in: "query"
        description: "Number of results per page"
        required: true
        type: "integer"
        minimum: 1
        maximum: 500
      responses:
        "200":
          description: "Successful operation"
        "401":
          description: "Unauthorized access"
        "403":
          description: "Forbidden"
        "500":
          description: "Server error"
      security:
      - v2_auth_query: []
      - v2_auth_header: []
  /v2/audits:
    get:
      tags:
      - "v2"
      summary: ""
      description: ""
      operationId: "getAuditsV2"
      produces:
      - "application/json"
      parameters:
      - name: "hkeys"
        in: "query"
        description: "Comma-separated list of hkeys"
        required: true
        type: "string"
      - name: "exclude"
        in: "query"
        description: "Comma-separated list of audit categories to be excluded from response. Audit categories are 'touchless','cleansafe','green', 'geosure' and 'checkin'."
        type: "string"
      - name: "include"
        in: "query"
        description: "Comma-separated list of audit categories to be included from response. Audit categories are 'touchless','cleansafe','green', 'geosure' and 'checkin'."
        type: "string"
      - name: "backfill"
        in: "query"
        description: "***For internal and educational purposes only - should not be used in public production environments.***\nOptional flag to provide calculated Green Stay record based on similar hotels, if no actual record exists for the hotel."
        required: false
        type: "boolean"
      responses:
        "200":
          description: "Successful operation"
        "401":
          description: "Unauthorized access"
        "403":
          description: "Forbidden"
        "500":
          description: "Server error"
      security:
      - v2_auth_query: []
      - v2_auth_header: []
    options:
      tags:
      - "v2"
      operationId: "corsGetAuditsV2"
      responses:
        '204':
          description: "Successful operation"
...
  /v2/gs_cert/{code}:
    get:
      tags:
      - "v2"
      summary: "Green Stay Certificate"
      description: "Green Stay Certificate"
      operationId: "getGreenStayCert"

      produces:
      - "text/html"
      parameters:
      - name: "code"
        in: "path"
        description: "Request code of certificate"
        type: "string"
        required: true
      responses:
        "200":
          description: "Successful operation"
        "403":
          description: "Forbidden"
        "500":
          description: "Server error"
  /v2/gs_widget/{type}/{code}:
    get:
      tags:
      - "v2"
      summary: "Green Stay Widget"
      description: "Green Stay Widget"
      operationId: "getGreenStayWidget"
      produces:
      - "image/png"
      parameters:
      - name: "type"
        in: "path"
        description: "Type of widget"
        type: "string"
        required: true
      - name: "code"
        in: "path"
        description: "Request code of widget"
        type: "string"
        required: true
      responses:
        "200":
          description: "Successful operation"
        "403":
          description: "Forbidden"
        "500":
          description: "Server error"
  /v2/gs_widget_link/{type}/{code}:
    get:
      tags:
      - "v2"
      summary: "Green Stay Widget Link Redirect"
      description: "Green Stay Link Redirect"
      operationId: "getGreenStayLink"
      parameters:
      - name: "type"
        in: "path"
        description: "Type of widget"
        type: "string"
        required: true
      - name: "code"
        in: "path"
        description: "Request code of widget"
        type: "string"
        required: true
      responses:
        "302":
          description: "Successful operation"
        "403":
          description: "Forbidden"
        "500":
          description: "Server error"
  /v2/gs_sticker/{type}/{code}:
    get:
      tags:
      - "v2"
      summary: "Green Stay Sticker"
      description: "Green Stay Sticker"
      operationId: "getGreenStaySticker"
      produces:
      - "application/pdf"
      parameters:
      - name: "type"
        in: "path"
        description: "Type of widget"
        type: "string"
        required: true
      - name: "code"
        in: "path"
        description: "Request code of widget"
        type: "string"
        required: true
      responses:
        "200":
          description: "Successful operation"
        "403":
          description: "Forbidden"
        "500":
          description: "Server error"
securityDefinitions:
  v1_auth:
    type: "apiKey"
    name: "secretKey"
    in: "query"
  v2_auth_query:
    type: "apiKey"
    name: "token"
    in: "query"
  v2_auth_header:
    type: "apiKey"
    name: "x-access-token"
    in: "header"
...